requirements.txt
fastapi
uvicorn
openai
python-dotenv
google-api-python-client
google-auth
